Separate hass's API calls from main program
